+++ Example/Tests/Tests.m
@@ -147,7 +147,7 @@ describe(@"Firebase Integration", ^{
 
         [integration track:payload];
         // TODO: look into how to handle mapping Firebase reserved params to each products
-        [verify(mockFirebase) logEventWithName:@"ecommerce_purchase" parameters:@{
+        [verify(mockFirebase) logEventWithName:@"purchase" parameters:@{
             @"checkout_id" : @"9bcf000000000000",
             @"transaction_id" : @"50314b8e",
             @"affiliation" : @"App Store",
@@ -156,11 +156,11 @@ describe(@"Firebase Integration", ^{
             @"tax" : @1.20,
             @"currency" : @"USD",
             @"item_category" : @"Games",
-            @"products" : @{
-                @"product_id" : @"2013294",
-                @"category" : @"Games",
-                @"name" : @"Monopoly: 3rd Edition",
-                @"brand" : @"Hasbros",
+            @"items" : @{
+                @"item_id" : @"2013294",
+                @"item_category" : @"Games",
+                @"item_name" : @"Monopoly: 3rd Edition",
+                @"item_brand" : @"Hasbros",
                 @"price" : @"21.99",
                 @"quantity" : @"1"
             }
@@ -192,7 +192,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"quantity" : @1,
@@ -229,7 +229,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"quantity" : @1,
@@ -268,7 +268,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"quantity" : @1,
@@ -306,7 +306,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"quantity" : @1,
@@ -351,11 +351,11 @@ describe(@"Firebase Integration", ^{
             @"tax" : @1.20,
             @"currency" : @"USD",
             @"item_category" : @"Games",
-            @"products" : @{
-                @"product_id" : @"2013294",
-                @"category" : @"Games",
-                @"name" : @"Monopoly: 3rd Edition",
-                @"brand" : @"Hasbros",
+            @"items" : @{
+                @"item_id" : @"2013294",
+                @"item_category" : @"Games",
+                @"item_name" : @"Monopoly: 3rd Edition",
+                @"item_brand" : @"Hasbros",
                 @"price" : @"21.99",
                 @"quantity" : @"1"
             }
@@ -444,11 +444,11 @@ describe(@"Firebase Integration", ^{
         [verify(mockFirebase) logEventWithName:@"view_item_list" parameters:@{
             @"list_id" : @"hot_deals_1",
             @"item_category" : @"Deals",
-            @"products" : @{
-                @"product_id" : @"2013294",
-                @"category" : @"Games",
-                @"name" : @"Monopoly: 3rd Edition",
-                @"brand" : @"Hasbros",
+            @"items" : @{
+                @"item_id" : @"2013294",
+                @"item_category" : @"Games",
+                @"item_name" : @"Monopoly: 3rd Edition",
+                @"item_brand" : @"Hasbros",
                 @"price" : @"21.99",
                 @"quantity" : @"1"
             }
@@ -485,7 +485,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"quantity" : @1,
@@ -525,7 +525,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"url" : @"https://www.company.com/product/path",
@@ -561,7 +561,7 @@ describe(@"Firebase Integration", ^{
             @"sku" : @"G-32",
             @"item_category" : @"Games",
             @"item_name" : @"Monopoly: 3rd Edition",
-            @"brand" : @"Hasbro",
+            @"item_brand" : @"Hasbro",
             @"variant" : @"200 pieces",
             @"price" : @18.99,
             @"url" : @"https://www.company.com/product/path",
@@ -585,11 +585,16 @@ describe(@"Firebase Integration", ^{
 
     it(@"track screen with name", ^{
         SEGScreenPayload *payload = [[SEGScreenPayload alloc] initWithName:@"Home screen"
+                                                                  category:@""
                                                                 properties:@{}
                                                                    context:@{}
                                                               integrations:@{}];
         [integration screen:payload];
-        [verify(mockFirebase) setScreenName:@"Home_screen" screenClass:nil];
+        // screen is set async, so need to pump the runloop.
+        [NSRunLoop.mainRunLoop runUntilDate:[NSDate dateWithTimeIntervalSinceNow:5]];
+        [verify(mockFirebase) logEventWithName:@"screen_view" parameters:@{
+            kFIRParameterScreenName : @"Home screen"
+        }];
     });
 
 });